refactor(oci): rename the docker format package to oci
