+++ internal/formats/oci/uploads.go
@@ -1,4 +1,4 @@
-package docker
+package oci
 
 import (
 	"bytes"
